# Optimize CI/CD: 3min faster builds with real quality gates