--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -107,18 +107,8 @@ jobs:
         echo ""
         echo "✅ Security audit complete"
 
-    - name: Verify Convex files are committed
-      run: |
-        echo "🔍 Checking required Convex files are present..."
-        for file in api.d.ts api.js dataModel.d.ts server.d.ts server.js; do
-          if [ ! -f "convex/_generated/$file" ]; then
-            echo "❌ Missing required file: convex/_generated/$file"
-            echo "These files MUST be committed to Git for deployments"
-            echo "Run 'npx convex codegen' and commit the files"
-            exit 1
-          fi
-        done
-        echo "✅ All required Convex files present"
+    - name: Verify Convex files
+      uses: ./.github/actions/verify-convex
       
     - name: Cache Convex generated files
       id: convex-cache
@@ -141,9 +131,6 @@ jobs:
     - name: Run type checking
       run: pnpm type-check
       
-    - name: Run tests
-      run: pnpm test
-
     - name: Run tests with coverage
       run: pnpm test:coverage
 
@@ -234,18 +221,8 @@ jobs:
           ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json', '**/pnpm-lock.yaml') }}-
           ${{ runner.os }}-nextjs-
       
-    - name: Verify Convex files are committed
-      run: |
-        echo "🔍 Checking required Convex files are present..."
-        for file in api.d.ts api.js dataModel.d.ts server.d.ts server.js; do
-          if [ ! -f "convex/_generated/$file" ]; then
-            echo "❌ Missing required file: convex/_generated/$file"
-            echo "These files MUST be committed to Git for deployments"
-            echo "Run 'npx convex codegen' and commit the files"
-            exit 1
-          fi
-        done
-        echo "✅ All required Convex files present"
+    - name: Verify Convex files
+      uses: ./.github/actions/verify-convex
       
     - name: Cache Convex generated files
       id: convex-cache