--- a/convex/_generated/api.d.ts
+++ b/convex/_generated/api.d.ts
@@ -8,11 +8,7 @@
  * @module
  */
 
-import type {
-  ApiFromModules,
-  FilterApi,
-  FunctionReference,
-} from "convex/server";
+import type { ApiFromModules, FilterApi, FunctionReference } from "convex/server";
 import type * as actions_historicalContext from "../actions/historicalContext.js";
 import type * as crons from "../crons.js";
 import type * as events from "../events.js";
@@ -40,11 +36,5 @@ declare const fullApi: ApiFromModules<{
   puzzles: typeof puzzles;
   users: typeof users;
 }>;
-export declare const api: FilterApi<
-  typeof fullApi,
-  FunctionReference<any, "public">
->;
-export declare const internal: FilterApi<
-  typeof fullApi,
-  FunctionReference<any, "internal">
->;
+export declare const api: FilterApi<typeof fullApi, FunctionReference<any, "public">>;
+export declare const internal: FilterApi<typeof fullApi, FunctionReference<any, "internal">>;