--- a/convex/_generated/dataModel.d.ts
+++ b/convex/_generated/dataModel.d.ts
@@ -26,10 +26,7 @@ export type TableNames = TableNamesInDataModel<DataModel>;
  *
  * @typeParam TableName - A string literal type of the table name (like "users").
  */
-export type Doc<TableName extends TableNames> = DocumentByName<
-  DataModel,
-  TableName
->;
+export type Doc<TableName extends TableNames> = DocumentByName<DataModel, TableName>;
 
 /**
  * An identifier for a document in Convex.
@@ -44,8 +41,7 @@ export type Doc<TableName extends TableNames> = DocumentByName<
  *
  * @typeParam TableName - A string literal type of the table name (like "users").
  */
-export type Id<TableName extends TableNames | SystemTableNames> =
-  GenericId<TableName>;
+export type Id<TableName extends TableNames | SystemTableNames> = GenericId<TableName>;
 
 /**
  * A type describing your Convex data model.