--- a/prettier.config.js
+++ b/prettier.config.js
@@ -1,0 +1,8 @@
+export default {
+  semi: true,
+  singleQuote: false,
+  tabWidth: 2,
+  trailingComma: "all",
+  printWidth: 100,
+  plugins: ["prettier-plugin-tailwindcss"],
+};