… .github/dependabot.yml
… .github/workflows/claude-code-review.yml
… .github/workflows/deploy.yml
… .github/workflows/size-limit.yml
… CLAUDE.md
… TODO.md
… package.json
… pnpm-lock.yaml
… scripts/manage-events.ts
… src/components/__tests__/NotificationModal.test.tsx